bg(168069761) Rename user can view mentors endpoint - rename localhost:8001/api/v1/mentor to localhost:8001/api/v1/mentors - modify all files which manipulate the endpoint [finishes #168069761]
