+++ server/src/v1/app.js
@@ -15,7 +15,7 @@ const port = process.env.PORT || 8001;
 app.use('/api/v1', createAdmin);
 app.use('/api/v1/auth', routes.authRoutes);
 app.use('/api/v1/user', routes.userRoutes);
-app.use('/api/v1/mentor', routes.mentorRoutes);
+app.use('/api/v1/mentors', routes.mentorsRoutes);
 
 app.listen(port, () => console.log(`Server started on port ${port}`));
